Add docs and updated flutter version
